Improve names for slate editor related modules.
