+++ packages/editor/src/client/blocks/richTextBlock/editor/utils.ts
