+++ packages/editor/src/client/blocks/richTextBlock/toolbar/tableMenu.tsx
@@ -3,9 +3,9 @@ import {useTranslation} from 'react-i18next'
 import {Button, InputGroup, InputNumber} from 'rsuite'
 import {Editor, Transforms, Element as SlateElement} from 'slate'
 import {useSlate} from 'slate-react'
-import {SubMenuContext} from '../../atoms/toolbar'
-import {isFormatActive} from './editorUtils'
-import {BlockFormat} from './formats'
+import {SubMenuContext} from '../../../atoms/toolbar'
+import {isFormatActive} from '../editor/utils'
+import {BlockFormat} from '../editor/formats'
 
 import './tableMenu.less'
